Add table column alignment option
Added ALIGN_TABLE_COLUMNS (= right, left, center) in the config.py file
and use it to specify the alignment in markdown.

Also fixed a problem where breaks in table cells (to create multiple
lines within the cell) got replaced by newlines, which broke the
table formatting in markdown.  If the <br> commands are retained
in the cell, that does the right thing in markdown too.

Finally, changed the entity substitution code in entityref() so
that less than, greater than symbols are not replaced by &lt; and &gt;
inside pre format tags.