Fix bug in tables when pad_tables is False
The initial and final vertical bars were missing when pad_tables was not
True.  This is probably a non-ideal fix for that -- it would be better
to make it work the same with or without pad_tables.  I didn't dig
deeply enough into the code to figure out how to do that.