Replace <p> with line break in list elements
The paragraph tag <p> is legal within list elements <li>, where it is
used to break the list item up into paragraphs.  The markdown equivalent
is to insert a break tag <br/> at the appropriate location.  This change
keeps track of when we are inside a list item and changes paragraph tags
to breaks.